Implement send coalescing for Process::send/GenServer::cast.
This implements an optimized path for sending messages to multiple, remote, processes. It will only send one packet to the node, with all of the targets. Currently, sending to multiple local processes is not as optimized as a loop. This will be solved once some form of specialization lands on Rust stable.